Refactor sidebar layout in VitePress configuration
The commit simplifies the previous nested structure of the sidebar on the website for easier maintenance. Now, the sidebar items are defined in separate functions and the new collapsible structure is used for all sections. The commit also includes minor changes such as reordering of import elements and reformatting texts and links.