Add new function for nth derivative
Introduced a new function in the math library for calculating nth order derivatives. The function takes three parameters: x, y, and n. The default value of n is 1, implying the first derivative. This new function will aid in performing iterative derivations and is more understandable than the traditional mathematical notation.